Cut RAM/disk footprint ~20x for data generation and checkpoints
- Image loaders store grayscale uint8 downscaled to the conversion
  canvas instead of full-size RGB float32 — the default 50k-image
  ImageNet-Sketch run drops from hundreds of GB of RAM to ~15 GB,
  with zero pipeline quality loss (the converter grayscales and
  letterboxes into the same canvas anyway)
- image_to_ascii_grid and _apply_segmentation accept the compact
  format alongside the original [3,H,W] float input
- Dataset files store grids as uint8 (98-char vocab fits in a byte):
  geometry 6.1 GB -> 0.77 GB, shading 3.1 GB -> 0.39 GB;
  ASCIIDataset casts to long at read time
- Rolling checkpoints (stage_last.pt + stage_best.pt) replace
  per-epoch files: ~12 GB -> ~2 GB over a full curriculum
- README documents concrete resource requirements

+++ data/prepare_human_ascii.py
@@ -137,7 +137,7 @@ def main():
         print("ERROR: No usable ASCII art found.")
         sys.exit(1)
 
-    data = torch.stack(grids)
+    data = torch.stack(grids).to(torch.uint8)  # vocab fits in a byte
     out_path = os.path.join(os.path.dirname(__file__), "human_data.pt")
     torch.save(data, out_path)
     print(f"Saved {data.shape} to {out_path} "
